LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load frame from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html' because of Content-Security-Policy.
